Updated source codes from cluster 11/19/2012
This brings in changes that were present on the cluster
in the /home/stuff/fengming/TEMmodules/DOSTEM1.2 folder
as of 11/19/2012. These changes allow for including 4
tundras and 3 boreal forests.
**NOTE**: The code may or may not run with the test files at
this point -- the various config files need to be updated.


Notes from Fengming's email:
So, please remember that: vegetation ID has been changed
as following:
0 – graminoid tundra;
1 – wet sedge tundra;
2 – shrub tundra;
3 – heather tundra;
4 – deciduous boreal forest;
5 – white spruce; and
6 – black spruce.
This implies that you need to regenerate the
"vegetation.nc" under cohort/eq/ for any cases.

And still there are two drainage types for each
vegetation types.